Fix board.nextColor. Make komi float.

+++ server/humanai/src/main/java/gotrainer/humanai/kata/Kata.java
@@ -55,7 +55,7 @@ public class Kata {
         clearReader();
     }
 
-    public void setKomi(int komi) throws Exception {
+    public void setKomi(float komi) throws Exception {
         write("komi " + komi);
         clearReader();
     }
